Refactor precision and recall classes
